+++ tests/deserialization/test_deserialize_series_list_from_toml.py
@@ -6,7 +6,7 @@
 import pytest
 
 from lanpartydb.deserialization import deserialize_series_list_from_toml
-from lanpartydb.models import Series
+from lanpartydb.models import Series, SeriesLinks, Resource
 
 
 @pytest.mark.parametrize(
@@ -44,6 +44,9 @@ from lanpartydb.models import Series
             title = "DeltaLAN"
             alternative_titles = ["Δ LAN", "Δέλτα LAN"]
             country_codes = ["au"]
+
+            [series.links.website]
+            url = "https://www.deltalan.example/"
             """,
             [
                 Series(
@@ -57,6 +60,12 @@ from lanpartydb.models import Series
                     title='DeltaLAN',
                     alternative_titles=['Δ LAN', 'Δέλτα LAN'],
                     country_codes=['au'],
+                    links=SeriesLinks(
+                        website=Resource(
+                            url='https://www.deltalan.example/',
+                            offline=False,
+                        ),
+                    ),
                 ),
             ],
         ),